+++ src/graphics/screens/crt_test_screen.cpp
@@ -50,7 +50,7 @@ CrtTestScreen::~CrtTestScreen() {}
 void CrtTestScreen::update(double delta_time)
 {
     display_test_dialog("CRT test", *screens, *screen_names, delta_time, *screen_manager);
-    sprite->transform.rotate(1.f * delta_time);
+    sprite->transform.rotate(1.f * static_cast<float>(delta_time));
 
     ImGui::Begin("Sprite position");
     float position[2]{sprite->transform.get_position().x, sprite->transform.get_position().y};
@@ -70,6 +70,6 @@ void CrtTestScreen::draw(Renderer &renderer, bool /*lightweight*/)
     crt_shader->set_uniform_2f("u_resolution", glm::vec2(renderer.get_width(), renderer.get_height()));
     crt_shader->set_uniform_1f("u_output_gamma", 1);
     crt_shader->set_uniform_1f("u_strength", 1);
-    crt_shader->set_uniform_1f("u_distortion_factor", 0.03);
+    crt_shader->set_uniform_1f("u_distortion_factor", 0.03f);
 }
 }
